refactor(app-router): share partial RSC shell state
PPR fallback shells owned the generic warmup and final render mechanics directly, which made later shell modes depend on PPR-specific state names.

Move the shared abort, dynamic-boundary, and cache-task lifecycle into a partial RSC shell primitive and keep the PPR module as a compatibility wrapper over that primitive.
